refactor(generation): split context_assembler into a context/ package
Break the 1013-line context_assembler.py into a context/ package:
- contexts.py     — the 10 per-template context dataclasses (+ _TopFile)
- token_budget.py — estimate_tokens / trim_to_budget / estimate_kg_tokens
                    as free functions, plus a new items_within_budget()
- graph_intelligence.py — the 3 stateless extract_* graph helpers
- assembler.py    — ContextAssembler + _symbol_to_dict

context_assembler.py becomes a façade re-exporting every public name, so
the existing imports (generation/__init__, page_generator/*, server, cli,
tests) are unchanged. ContextAssembler's _estimate_tokens / _trim_to_budget
/ _estimate_kg_tokens are now thin wrappers delegating to token_budget, so
no internal call site changed. The two symbol-selection budget loops in
assemble_file_page collapse onto items_within_budget().

No behavior change. Ruff clean, no new mypy errors; tests/unit green
(2383 passed).
